pyproject.toml
[tool.poetry]
name = "fit-ctf"
version = "0.1.0"
description = "A CTF backend for an assignment of a class BIS at FIT, Brno University of Technology."
authors = ["Hung Do <hungdojan@gmail.com>"]
readme = "README.md"
repository = "https://github.com/hungdojan/FIT_CTF"
packages = [
    { include = "fit_ctf_backend", from = "src" },
    { include = "fit_ctf_rendezvous", from = "src" },
    { include = "fit_ctf_models", from = "src" },
    { include = "fit_ctf_templates", from = "src"},
    { include = "fit_ctf_utils", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.3"
passlib = "^1.7.4"
pymongo = "^4.6.2"
redis = "^5.0.3"
pytermgui = "^7.7.1"
python-dotenv = "^1.0.1"
click = "^8.1.7"
tabulate = "^0.9.0"


setuptools = "^70.0.0"
jsonschema = "^4.23.0"
pydantic = "^2.10.6"
[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
coverage = "^7.4.3"
hypothesis = "^6.98.11"
pytest-cov = "^4.1.0"
pytest-html = "^4.1.1"
pytest-mock = "^3.12.0"
pytest-picked = "^0.5.0"
pytest-sugar = "^1.0.0"


pyyaml = "^6.0.1"
pytest-env = "^1.1.3"
[tool.poetry.group.docs.dependencies]
sphinx = "^7.3.7"

click-man = "^0.4.1"
sphinx-click = "^6.0.0"
sphinx-rtd-theme = "^2.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"
isort = "^6.0.0"
flake8-pyproject = "^1.2.3"
pre-commit = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
env = [
    "CONTAINER_CLIENT=mock",
    "DB_CONNECTION_TIMEOUT=5",
    "DB_NAME=test-ctf-db",
    "ENABLE_CONTAINER_CLIENT_TESTING=0"
]
addopts = "-vs"

[tool.coverage.html]
directory = "tests/reports/htmlcov"

[tool.coverage.run]
omit = [
    "tests/*",
    "/tmp/*",
    "src/fit_ctf_utils/*.py",
    "src/fit_ctf_utils/container_client/*",
    "**/__init__.py",
]
command_line = "-m pytest -vv"

[tool.flake8]
max-line-length = 100
extend-exclude = [
    ".venv", "docs"
]
per-file-ignores = [
    "**/__init__.py:F4"
]
extend-ignore = ["E226", "E266", "W503"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fit-ctf = "fit_ctf_backend.__main__:main"
ctf-rendezvous = "fit_ctf_rendezvous.__main__:main"
